コードの構成を整理してインターフェイスで集約する
